フロントの検索部分->laravel側のテンプレート修正及び詳細ページを作成する

+++ database/migrations/2021_11_12_201932_create_commentables_table.php
@@ -14,7 +14,7 @@ class CreateCommentablesTable extends Migration
     public function up()
     {
         Schema::create('commentables', function (Blueprint $table) {
-            $table->bigIncrements('comment_id');
+            $table->integer('comment_id');
             $table->integer('comentable_id');
             $table->String('commentable_type');
             $table->timestamps();
